refactor: split DuckDB config section into config_duckdb.go
Move DuckDBConfig, RoutingPolicy, RoutingStrategy types and
defaultDuckDBConfig() to config_duckdb.go. Extract DuckDB-specific
validation from Validate() into validateDuckDBConfig().

config.go reduced from 600 to 486 lines (under 500-line limit).
No exported symbols changed, no behavioral changes.

Closes #136
